requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.11.0
    # via
    #   -r requirements.in
    #   django
dj-database-url==3.0.1
    # via -r requirements.in
django==6.0
    # via
    #   -r requirements.in
    #   dj-database-url
    #   gateway-defender
gateway-defender==0.1.3
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
packaging==25.0
    # via
    #   -r requirements.in
    #   gunicorn
psycopg2-binary==2.9.11
    # via -r requirements.in
python-decouple==3.8
    # via -r requirements.in
sqlparse==0.5.4
    # via
    #   -r requirements.in
    #   django
typing-extensions==4.15.0
    # via -r requirements.in
whitenoise==6.11.0
    # via -r requirements.in
